refactor: 이미지 파일 변경 및 파일 정리
