+++ frontend/src/assets/index.ts
@@ -1,4 +1,9 @@
-export { default as profile } from "./images/profile.svg";
+// 아이콘
 export { default as tab } from "./icons/tab.svg";
 export { default as enter } from "./icons/enter.svg";
 export { default as spacebar } from "./icons/spacebar.svg";
+export { default as backIcon } from "./icons/back.svg";
+
+// 이미지
+export { default as logoImg } from "./images/CodeWith.svg";
+export { default as profile } from "./images/profile.svg";
